Move all the existing interactors to Domains namespace
