--- app/interactions/cancel_force_delete_interaction/base.rb
@@ -1,7 +1,0 @@
-module CancelForceDeleteInteraction
-  class Base < ActiveInteraction::Base
-    object :domain,
-           class: Domain,
-           description: 'Domain to cancel ForceDelete on'
-  end
-end
