--- app/interactions/cancel_force_delete_interaction/cancel_force_delete.rb
@@ -1,10 +1,0 @@
-module CancelForceDeleteInteraction
-  class CancelForceDelete < Base
-    def execute
-      compose(RemoveForceDeleteStatuses, inputs)
-      compose(RestoreStatusesBeforeForceDelete, inputs)
-      compose(ClearForceDeleteData, inputs)
-      compose(NotifyRegistrar, inputs)
-    end
-  end
-end
