--- app/interactions/cancel_force_delete_interaction/clear_force_delete_data.rb
@@ -1,10 +1,0 @@
-module CancelForceDeleteInteraction
-  class ClearForceDeleteData < Base
-    def execute
-      domain.force_delete_data = nil
-      domain.force_delete_date = nil
-      domain.force_delete_start = nil
-      domain.save(validate: false)
-    end
-  end
-end
