--- app/interactions/cancel_force_delete_interaction/notify_registrar.rb
@@ -1,8 +1,0 @@
-module CancelForceDeleteInteraction
-  class NotifyRegistrar < Base
-    def execute
-      domain.registrar.notifications.create!(text: I18n.t('force_delete_cancelled',
-                                                          domain_name: domain.name))
-    end
-  end
-end
