--- app/interactions/cancel_force_delete_interaction/remove_force_delete_statuses.rb
@@ -1,11 +1,0 @@
-module CancelForceDeleteInteraction
-  class RemoveForceDeleteStatuses < Base
-    def execute
-      domain.statuses.delete(DomainStatus::FORCE_DELETE)
-      domain.statuses.delete(DomainStatus::SERVER_RENEW_PROHIBITED)
-      domain.statuses.delete(DomainStatus::SERVER_TRANSFER_PROHIBITED)
-      domain.statuses.delete(DomainStatus::CLIENT_HOLD)
-      domain.save(validate: false)
-    end
-  end
-end
