--- app/interactions/cancel_force_delete_interaction/restore_statuses_before_force_delete.rb
@@ -1,9 +1,0 @@
-module CancelForceDeleteInteraction
-  class RestoreStatusesBeforeForceDelete < Base
-    def execute
-      domain.statuses = domain.statuses_before_force_delete
-      domain.statuses_before_force_delete = nil
-      domain.save(validate: false)
-    end
-  end
-end
