+++ app/interactions/domains/cancel_force_delete/base.rb
@@ -1,0 +1,9 @@
+module Domains
+  module CancelForceDelete
+    class Base < ActiveInteraction::Base
+      object :domain,
+             class: Domain,
+             description: 'Domain to cancel ForceDelete on'
+    end
+  end
+end
