+++ app/interactions/domains/cancel_force_delete/cancel_force_delete.rb
@@ -1,0 +1,12 @@
+module Domains
+  module CancelForceDelete
+    class CancelForceDelete < Base
+      def execute
+        compose(RemoveForceDeleteStatuses, inputs)
+        compose(RestoreStatusesBeforeForceDelete, inputs)
+        compose(ClearForceDeleteData, inputs)
+        compose(NotifyRegistrar, inputs)
+      end
+    end
+  end
+end
