+++ app/interactions/domains/cancel_force_delete/notify_registrar.rb
@@ -1,0 +1,10 @@
+module Domains
+  module CancelForceDelete
+    class NotifyRegistrar < Base
+      def execute
+        domain.registrar.notifications.create!(text: I18n.t('force_delete_cancelled',
+                                                            domain_name: domain.name))
+      end
+    end
+  end
+end
