+++ app/interactions/domains/cancel_force_delete/restore_statuses_before_force_delete.rb
@@ -1,0 +1,11 @@
+module Domains
+  module CancelForceDelete
+    class RestoreStatusesBeforeForceDelete < Base
+      def execute
+        domain.statuses = domain.statuses_before_force_delete
+        domain.statuses_before_force_delete = nil
+        domain.save(validate: false)
+      end
+    end
+  end
+end
