+++ app/interactions/domains/force_delete/check_discarded.rb
@@ -1,0 +1,12 @@
+module Domains
+  module ForceDelete
+    class CheckDiscarded < Base
+      def execute
+        return true unless domain.discarded?
+
+        message = 'Force delete procedure cannot be scheduled while a domain is discarded'
+        errors.add(:domain, message)
+      end
+    end
+  end
+end
