--- app/interactions/force_delete_interaction/prepare_domain.rb
@@ -1,13 +1,0 @@
-module ForceDeleteInteraction
-  class PrepareDomain < Base
-    STATUSES_TO_SET = [DomainStatus::FORCE_DELETE,
-                       DomainStatus::SERVER_RENEW_PROHIBITED,
-                       DomainStatus::SERVER_TRANSFER_PROHIBITED].freeze
-
-    def execute
-      domain.statuses_before_force_delete = domain.statuses
-      domain.statuses |= STATUSES_TO_SET
-      domain.save(validate: false)
-    end
-  end
-end
